address NbootCtx inputs on an unprovisioned MCU
