+++ examples/mcxa-577app/bootloader/.cargo/config.toml
@@ -2,7 +2,7 @@
 target = "thumbv8m.main-none-eabihf"
 
 [target.thumbv8m.main-none-eabihf]
-runner = "probe-rs run --chip MCXA577 --chip-description-path MCXA_custom.yaml"
+runner = "probe-rs run --chip MCXA577 --speed 10000"
 
 [target.'cfg(all(target_arch = "arm", target_os = "none"))']
 rustflags = [
