+++ examples/mcxa-577app/bootloader/memory.x
@@ -6,9 +6,9 @@ MEMORY
    */
   FLASH (rx) : ORIGIN = 0x00000000, LENGTH = 64K
 
-  /* Secure SRAM alias: 0x3000_0000 (Matrix0 Target Port 4, Secure, All Initiators).
+  /* Use the same SRAM base as the working direct-run apps so RTT is discoverable.
    */
-  RAM (rwx) : ORIGIN = 0x30000000, LENGTH = 64K
+  RAM (rwx) : ORIGIN = 0x20000000, LENGTH = 64K
 }
 
 /* Stack grows down from end of RAM */
